feat(ledger): NUMERIC(18,2) money columns + unenumerable Paynow refs
Two production-readiness items the operator-DD memo flagged as shippable
before any real money touches the system.

NUMERIC(18,2) migration
  accounts.balance, transactions.amount, transactions.fee,
  ledger_entries.amount, ledger_entries.balance_after — all move from
  DOUBLE PRECISION to NUMERIC(18,2). Binary float arithmetic silently
  drifts on decimal values (0.1 + 0.2 != 0.3); that error compounds
  across millions of ledger entries into a silent-but-growing delta
  between balance and ledger_net. NUMERIC is exact decimal arithmetic
  with enough headroom (10^16 USD) for anything plausible.

  Models now return Decimal; payments.py has a _to_decimal() boundary
  helper so float amounts from Pydantic get cast cleanly (via
  Decimal(str(x)) to avoid binary-repr noise) before any math. Paynow
  SDK calls get float(amount) at the wire boundary since the SDK
  expects primitives.

  scripts/migrate_money_columns_to_numeric.py is a one-shot Postgres
  migration for environments created before the switch — runs each
  ALTER in a single transaction with pre/post sum checks and a
  per-row tolerance, rolls back if drift exceeds tolerance.

Paynow reference hardening
  The old format `ZIPPIE-{tx_id}` was enumerable: seeing one reference
  (in logs, support tickets, or Paynow dashboards) leaked the sequence
  of adjacent transactions. New format `ZIPPIE-{uuid12}-{tx_id}` keeps
  the tx_id suffix so webhook lookup stays O(1), but prefixes an 12-hex
  UUID segment that makes references unguessable. _parse_tx_id_from_reference
  accepts both formats so in-flight legacy references keep resolving.

Full suite: 28 passed (all existing, incl. 50-thread concurrency test).

+++ backend/app/db/models.py
@@ -8,7 +8,7 @@ from sqlalchemy import (
     CheckConstraint,
     Column,
     DateTime,
-    Float,
+    Numeric,
     ForeignKey,
     Index,
     Integer,
@@ -54,7 +54,7 @@ class Account(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     name = Column(String, nullable=False)
-    balance = Column(Float, default=0.0, nullable=False)
+    balance = Column(Numeric(18, 2), default=0, nullable=False)
     currency = Column(String, default="USD", nullable=False)  # USD, ZWL
     account_type = Column(String, default="primary")  # primary, savings, investment
     color = Column(String, default="#10b981")
@@ -76,7 +76,7 @@ class Transaction(Base):
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     account_id = Column(Integer, ForeignKey("accounts.id"), nullable=True)
     transaction_type = Column(String, nullable=False)  # sent, received, request
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(18, 2), nullable=False)
     currency = Column(String, default="USD", nullable=False)
     recipient = Column(String, nullable=False, index=True)
     sender = Column(String, nullable=True, index=True)
@@ -85,7 +85,7 @@ class Transaction(Base):
         String, default="pending", nullable=False
     )  # completed, pending, failed
     payment_method = Column(String, nullable=True)
-    fee = Column(Float, default=0.0)
+    fee = Column(Numeric(18, 2), default=0)
     transaction_metadata = Column(JSON, nullable=True)  # Additional data
     created_at = Column(DateTime(timezone=True), server_default=func.now(), index=True)
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
@@ -121,9 +121,9 @@ class LedgerEntry(Base):
     account_id = Column(
         Integer, ForeignKey("accounts.id"), nullable=False, index=True
     )
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(18, 2), nullable=False)
     direction = Column(String, nullable=False)  # 'debit' or 'credit'
-    balance_after = Column(Float, nullable=False)
+    balance_after = Column(Numeric(18, 2), nullable=False)
     created_at = Column(
         DateTime(timezone=True), server_default=func.now(), index=True
     )
